improved package structure
