+++ code/lib/metrics/__init__.py
@@ -1,5 +1,5 @@
 from .MR1 import *
 from .mAP import *
 from .mAP_at_15 import *
-from .utils import * 
+from .relevance import *
